pyproject.toml
[project]
name = "inference-core"
version = "0.2.0"
description = "A ready-to-use FastAPI backend with Celery-powered background tasks and first-class LLM integration using LangChain."
authors = [
    {name = "Tomasz Osiński"}
]
license = {text = "MIT"}
readme = "README.md"
requires-python = ">=3.12,<4.0"
dependencies = [
    "fastapi[standard] (>=0.116.1,<0.117.0)",
    "uvicorn[standard]",
    "pydantic-settings (>=2.10.1,<3.0.0)",
    "python-json-logger (>=3.3.0,<4.0.0)",
    "sentry-sdk[fastapi] (>=2.34.1,<3.0.0)",
    "sqlalchemy (>=2.0.42,<3.0.0)",
    "aiomysql (>=0.2.0,<0.3.0)",
    "cryptography (>=45.0.5,<46.0.0)",
    "aiosqlite (>=0.21.0,<0.22.0)",
    "asyncpg (>=0.30.0,<0.31.0)",
    "psycopg[binary] (>=3.1.18,<4.0.0)",
    "pymysql (>=1.1.1,<2.0.0)",
    "celery[redis] (>=5.5.3,<6.0.0)",
    "flower (>=2.0.1,<3.0.0)",
    "langchain-openai (>=0.3.29,<0.4.0)",
    "langchain-community (>=0.3.27,<0.4.0)",
    "python-jose[cryptography] (>=3.3.0,<4.0.0)",
    "passlib[bcrypt] (>=1.7.4,<2.0.0)",
    "bcrypt (>=4.3.0,<5.0.0)",
    "redis (>=5.0.0,<6.0.0)",
    "langchain-google-genai (>=2.1.9,<3.0.0)",
    "langchain-anthropic (>=0.3.18,<0.4.0)",
    "google-genai (>=1.30.0,<2.0.0)",
    "anthropic (>=0.64.0,<0.65.0)",
    "prometheus-client (>=0.21.1,<0.22.0)",
]

[tool.poetry]
packages = [
    { include = "inference_core" },
]

[tool.poetry.group.dev.dependencies]
pytest = "^8.4.1"
pytest-asyncio = "^1.1.0"
pytest-cov = "^6.2.1"
debugpy = "^1.8.16"
locust = "^2.38.0"

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
